move DockerContainer to utils
